+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/meta/MetaReader.kt
@@ -2,7 +2,7 @@ package space.kscience.dataforge.meta
 
 import space.kscience.dataforge.meta.descriptors.Described
 
-public interface MetaSpec<out T> : Described {
+public interface MetaReader<out T> : Described {
 
     /**
      * Read the source meta into an object and return null if Meta could not be interpreted as a target type
@@ -10,12 +10,12 @@ public interface MetaSpec<out T> : Described {
     public fun readOrNull(source: Meta): T?
 
     /**
-     * Read generic read-only meta with this [MetaSpec] producing instance of the desired type.
+     * Read generic read-only meta with this [MetaReader] producing instance of the desired type.
      * Throws an error if conversion could not be done.
      */
     public fun read(source: Meta): T = readOrNull(source) ?: error("Meta $source could not be interpreted by $this")
 }
 
 
-public fun <T : Any> MetaSpec<T>.readNullable(item: Meta?): T? = item?.let { read(it) }
-public fun <T> MetaSpec<T>.readValue(value: Value): T? = read(Meta(value))
+public fun <T : Any> MetaReader<T>.readNullable(item: Meta?): T? = item?.let { read(it) }
+public fun <T> MetaReader<T>.readValue(value: Value): T? = read(Meta(value))