+++ dataforge-meta/src/commonMain/kotlin/space/kscience/dataforge/misc/cast.kt
@@ -1,3 +1,0 @@
-package space.kscience.dataforge.misc
-
-public expect inline fun <T> Any?.unsafeCast(): T