+++ dataforge-meta/src/nativeMain/kotlin/space/kscience/dataforge/misc/castNative.kt
@@ -1,4 +1,0 @@
-package space.kscience.dataforge.misc
-
-@Suppress("UNCHECKED_CAST")
-public actual inline fun <T> Any?.unsafeCast(): T = this as T